+++ lib/presentation/player_round/player_round_modal.dart
@@ -98,6 +98,7 @@ class _PlayerRoundModalState extends ConsumerState<PlayerRoundModal> {
           scoreFilter: widget.scoreFilter,
           autofocus: widget.gameMode != GameMode.frenchDriving,
           enabled: widget.gameMode != GameMode.frenchDriving,
+          allowNegative: widget.gameMode == GameMode.standard,
           onSubmitted: _shouldCloseOnReturnKey() ? _closeModal : null,
         ),
       ],